requirements.txt
streamlit>=1.28.0
litellm>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
youtube-transcript-api>=1.2.0
requests>=2.31.0
reportlab>=4.0.0
markdown>=3.5.0
